LayoutTests/platform/mac/accessibility/frame-with-title.html

<!DOCTYPE HTML PUBLIC "-//IETF//DTD HTML//EN">
<html>
<head>
<script src="../../../fast/js/resources/js-test-pre.js"></script>
</head>
<body id="body">

<iframe id="frame1" src="image-map2.html" width=100 height=100 name="packageListFrame" title="All Packages">
</iframe>

<iframe id="frame2" src="image-map2.html" width=100 height=100 name="packageListFrame" >
</iframe>

<p id="description"></p>
<div id="console"></div>

<script>

    description("This test makes sure that a frame with a title will use that value instead of the name.");

    if (window.accessibilityController) {

          document.getElementById("frame1").focus();
          shouldBe("accessibilityController.focusedElement.description", "'AXDescription: All Packages'");

          document.getElementById("frame2").focus();
          shouldBe("accessibilityController.focusedElement.description", "'AXDescription: packageListFrame'");
    }

</script>

<script src="../../../fast/js/resources/js-test-post.js"></script>
</body>
</html>
